Move classes from RX to our package - This should prevent problems with conflicts in projects which use both Reactive Kafka and Kafka RX
