+++ src/main/scala/com/softwaremill/react/kafka/commit/OffsetCommitter.scala
@@ -1,0 +1,12 @@
+package com.softwaremill.react.kafka.commit
+
+/**
+ * Based on from https://github.com/cjdev/kafka-rx
+ */
+trait OffsetCommitter {
+  def commit(offsets: OffsetMap): OffsetMap
+
+  // optional / default fns below
+  def start(): Unit = ()
+  def stop(): Unit = ()
+}
