+++ src/main/scala/com/softwaremill/react/kafka/commit/PartitionLock.scala
@@ -1,0 +1,9 @@
+package com.softwaremill.react.kafka.commit
+
+/**
+ * Based on from https://github.com/cjdev/kafka-rx
+ */
+trait PartitionLock {
+  def acquire(): Unit
+  def release(): Unit
+}
